Rename signed and unsigned integer interface.
